Enable global error checking (`set -e`)

+++ buildsys/bamboo.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 # bamboo.sh
 
+set -eo pipefail
+
 # Description:
 
 # A shell script to command a build ala Jenkins
@@ -19,7 +21,6 @@ pwd
 df -h .
 echo ' '
 
-## export SST_BUILDOUTOFSOURCE=1
 #-------------------------------------------------------------------------
 # Function: TimeoutEx
 # Description:
@@ -32,9 +33,8 @@ echo ' '
 #   Return value: The return value of the command being run or !=0 to indicate
 #   a timeout or error.
 TimeoutEx() {
-    # Call (via "source") the moduleex.sh script with the passed in parameters
-    $SST_ROOT/../sqe/test/utilities/TimeoutEx.sh $@
-    # Get the return value from the moduleex.sh
+    local retval=0
+    $SST_ROOT/../sqe/test/utilities/TimeoutEx.sh $@ || retval=$?
     return $retval
 }
 
@@ -43,7 +43,6 @@ cloneOtherRepos() {
 ##  Check out other repositories except second time on Make Dist test
 
 if [ ! -d ../../distTestDir ] ; then
-## if [[ ${SST_TEST_ROOT:+isSet} != isSet ]] ; then
     echo "PWD $LINENO = `pwd`"
 
 ## Set the clone depth parameter
@@ -461,10 +460,9 @@ dotests() {
 #   Output: Any output from the module command.
 #   Return value: 0 on success, On error, bamboo.sh will exit with the moduleex.sh error code.
 ModuleEx() {
+    local retval=0
     # Call (via "source") the moduleex.sh script with the passed in parameters
-    . $SST_ROOT/test/utilities/moduleex.sh $@
-    # Get the return value from the moduleex.sh
-    retval=$?
+    source $SST_ROOT/test/utilities/moduleex.sh $@ || retval=$?
     if [ $retval -ne 0 ] ; then
         echo "ERROR: 'module' failed via script $SST_ROOT/test/utilities/moduleex.sh with retval= $retval; bamboo.sh exiting"
         exit $retval
@@ -938,17 +936,13 @@ linuxSetMPI() {
       echo "Attempt to initialize the modules utility.  Look for modules init file in 1 of 2 places"
 
       echo "Location 1: ls -l /etc/profile.modules"
-      ls -l /etc/profile.modules
-      if [ -f /etc/profile.modules ] ; then
-          . /etc/profile.modules
+      echo "Location 2: ls -l /etc/profile.d/modules.sh"
+      if [ -r /etc/profile.modules ] ; then
+          source /etc/profile.modules
           echo "bamboo.sh: loaded /etc/profile.modules"
-      else
-          echo "Location 2: ls -l /etc/profile.d/modules.sh"
-          ls -l /etc/profile.d/modules.sh
-          if [ -r /etc/profile.d/modules.sh ] ; then
-              source /etc/profile.d/modules.sh
-              echo "bamboo.sh: loaded /etc/profile.d/modules"
-          fi
+      elif [ -r /etc/profile.d/modules.sh ] ; then
+          source /etc/profile.d/modules.sh
+          echo "bamboo.sh: loaded /etc/profile.d/modules"
       fi
    fi
 
@@ -1019,11 +1013,10 @@ linuxSetMPI() {
         echo "bamboo.sh: Load GLPK"
         # Load available GLPK, whatever version it is
         ModuleEx load glpk
-       
+
         # METIS 5.1.0
         echo "bamboo.sh: Load METIS 5.1.0"
-        ModuleEx avail | grep bundled
-        if [ $? == 0 ] ; then
+        if ModuleEx avail | grep bundled ; then
             echo " Bingo ###################################################"
             ModuleEx load metis/metis-5.1.0-bundled
         else
@@ -1034,16 +1027,14 @@ linuxSetMPI() {
 
     else # otherwise try to load compiler-specific tool variant
         # GNU Linear Programming Kit (GLPK)
-        ModuleEx avail | egrep -q "glpk/glpk-4.54_${compiler}"
-        if [ $? == 0 ] ; then
+        if ModuleEx avail | grep -E -q "glpk/glpk-4.54_${compiler}" ; then
             echo "bamboo.sh: Load GLPK (gcc ${compiler} variant)"
             ModuleEx load glpk/glpk-4.54_${compiler}
         else
             echo "bamboo.sh: module GLPK (gcc ${compiler} variant) Not Available"
         fi
         # METIS 5.1.0
-        ModuleEx avail | egrep -q "metis/metis-5.1.0_${compiler}"
-        if [ $? == 0 ] ; then
+        if ModuleEx avail | grep -E -q "metis/metis-5.1.0_${compiler}" ; then
             if [[ ${compiler} != *intel-15* ]] ; then
                 echo "bamboo.sh: Load METIS 5.1.0 (gcc ${compiler} variant)"
                 ModuleEx load metis/metis-5.1.0_${compiler}
@@ -1068,7 +1059,7 @@ linuxSetMPI() {
 #                 Also need $2 passed along
 
 ldModules_MacOS_Clang() {
-    ClangVersion=$1            #   example "clang-700.0.72" $2
+    local ClangVersion=$1            #   example "clang-700.0.72" $2
     ModuleEx avail
     # Use MPI built with CLANG from Xcode
     ModuleEx unload mpi
@@ -1077,7 +1068,7 @@ ldModules_MacOS_Clang() {
     # GNU Linear Programming Kit (GLPK)
     echo "bamboo.sh: Load GLPK"
     ModuleEx load glpk/glpk-4.54_$ClangVersion
-        
+
     # METIS 5.1.0
     echo "bamboo.sh: Load METIS 5.1.0"
     ModuleEx load metis/metis-5.1.0_$ClangVersion
@@ -1109,8 +1100,8 @@ ldModules_MacOS_Clang() {
             ;;
     esac
 
-    export CC=`which clang`
-    export CXX=`which clang++`
+    export CC="$(command -v clang)"
+    export CXX="$(command -v clang++)"
     echo "    Modules loaded"
     ModuleEx list
     $CC --version
@@ -1144,40 +1135,26 @@ darwinSetMPI() {
     # Initialize modules for Jenkins (taken from $HOME/.bashrc on Mac)
     if [ -f /etc/profile.modules ]
     then
-        . /etc/profile.modules
-        echo "bamboo.sh: loaded /etc/profile.modules. Available modules"
-        ModuleEx avail
+        source /etc/profile.modules
+        echo "bamboo.sh: loaded /etc/profile.modules."
         # put any module loads here
         echo "bamboo.sh: Loading Modules for MacOSX"
-        # Do things specific to the MacOS version
         case $macosVersion in
-################################################################################
-
-            11.6) # Big Sur
+            11.6)
                 echo    "This is Big Sur, Compiler is $compiler"
-                ldModules_MacOS_Clang $compiler  $2   # any Xcode
                 ;;
-
-################################################################################
-
-            12.3) # Monterey
+            12.3)
                 echo    "This is Monterey, Compiler is $compiler"
-                ldModules_MacOS_Clang $compiler  $2   # any Xcode
                 ;;
-
-################################################################################
-            13.2) # Ventura
+            13.2)
                 echo    "This is Ventura, Compiler is $compiler"
-                ldModules_MacOS_Clang $compiler $2 # any code
                 ;;
-
-################################################################################
-            *) # unknown
+            *)
                 echo "bamboo.sh: Unknown Mac OS version. $macosVersion"
-                echo ' '
                 exit
                 ;;
         esac
+        ldModules_MacOS_Clang $compiler $2  # any Xcode
 
     else
         echo "ERROR: unable to locate /etc/profile.modules - cannot load modules"
@@ -1199,7 +1176,7 @@ darwinSetMPI() {
 setUPforMakeDisttest() {
     echo "Setting up to build from the tars created by make dist"
     echo "---   PWD $LINENO  `pwd`"           ## Original trunk
-     
+
     LOC_OF_TAR=""
     if [[ ${SST_BUILDOUTOFSOURCE:+isSet} == isSet ]] ; then
         LOC_OF_TAR="-builddir"
@@ -1420,6 +1397,16 @@ setUPforMakeDisttest() {
      fi
 }         #### End of setUPforMakeDistest  ####
 
+print_and_dump_loc() {
+    local loc="${1}"
+    if [[ -r "${loc}" ]]; then
+        echo "cat ${loc}"
+        cat "${loc}"
+    else
+        echo "not found: ${loc}"
+    fi
+}
+
 #-------------------------------------------------------------------------
 # Function: dobuild
 # Description:
@@ -1480,7 +1467,7 @@ dobuild() {
     # Mac OS X needs some help finding dylibs
     if [ $kernel == "Darwin" ]
     then
-	    export DYLD_LIBRARY_PATH=${LD_LIBRARY_PATH}:${DYLD_LIBRARY_PATH}
+		export DYLD_LIBRARY_PATH=${LD_LIBRARY_PATH}:${DYLD_LIBRARY_PATH}
     fi
 
     # Dump pre-build environment and modules status
@@ -1635,20 +1622,17 @@ dobuild() {
             echo "+++++++++++++++++++++++++++++++++++++++++++++++++++ makeDist"
         fi
 
+        [[ $kernel == "Darwin" ]] && linkage_display="otool -L" || linkage_display=ldd
+
         echo "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
         echo ' '
         echo "bamboo.sh: make on SST-CORE"
         echo ' '
         echo "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
 
-        if [ $kernel == "Darwin" ]
-        then
-            # Mac OS X
-            echo "$ otool -L $coresourcedir/src/sst/core/sstsim.x"
-            otool -L $coresourcedir/src/sst/core/sstsim.x
-        else
-            echo "$ ldd $coresourcedir/src/sst/core/sstsim.x"
-            ldd $coresourcedir/src/sst/core/sstsim.x
+        echo "${linkage_display} $coresourcedir/src/sst/core/sstsim.x"
+        if [[ -r $coresourcedir/src/sst/core/sstsim.x ]]; then
+            ${linkage_display} $coresourcedir/src/sst/core/sstsim.x
         fi
         echo "SST-CORE BUILD INFO============================================================"
 
@@ -1678,13 +1662,12 @@ dobuild() {
         echo ' '
         echo "bamboo.sh: make clean on SST-CORE"
         echo ' '
-        echo "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"        
+        echo "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
         make clean
 
         echo
         echo "=== DUMPING The SST-CORE installed sstsimulator.conf file ==="
-        echo "cat $SST_CORE_INSTALL/etc/sst/sstsimulator.conf"
-        cat $SST_CORE_INSTALL/etc/sst/sstsimulator.conf
+        print_and_dump_loc $SST_CORE_INSTALL/etc/sst/sstsimulator.conf
         echo "=== DONE DUMPING ==="
         echo
 
@@ -1695,14 +1678,10 @@ dobuild() {
         echo "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
         echo " "
 
-        if [ $kernel == "Darwin" ]
-        then
-            # Mac OS X
-            echo "$ otool -L $coresourcedir/src/sst/core/sstsim.x"
-            otool -L $coresourcedir/src/sst/core/sstsim.x
-        else
-            echo "$ ldd $coresourcedir/src/sst/core/sstsim.x"
-            ldd $coresourcedir/src/sst/core/sstsim.x
+        # TODO either this one or the previous one has the wrong path
+        echo "${linkage_display} $coresourcedir/src/sst/core/sstsim.x"
+        if [[ -r $coresourcedir/src/sst/core/sstsim.x ]]; then
+            ${linkage_display} $coresourcedir/src/sst/core/sstsim.x
         fi
         echo "SST-CORE BUILD INFO============================================="
 
@@ -1819,11 +1798,7 @@ if [[ $SST_SELECTED_ELEMENTS_CONFIG == "NOBUILD" ]]
         ls -ltrd * | tail -20
 
         echo "################################## DEBUG DATA ########################"
-        ls
-        ls src
-        ls src/sst
-        ls src/sst/elements/
-        ls src/sst/elements/*/*m4
+        find "$PWD" -type f | sort
         echo "##################### END ######## DEBUG DATA ########################"
 
         # Check to see if we are actually performing make dist
@@ -1901,15 +1876,13 @@ if [[ $SST_SELECTED_ELEMENTS_CONFIG == "NOBUILD" ]]
 
             echo
             echo "=== DUMPING The SST-ELEMENTS installed $HOME/.sst/sstsimulator.conf file ==="
-            echo "cat $HOME/.sst/sstsimulator.conf"
-            cat $HOME/.sst/sstsimulator.conf
+            print_and_dump_loc $HOME/.sst/sstsimulator.conf
             echo "=== DONE DUMPING ==="
             echo
 
             echo
             echo "=== DUMPING The SST-ELEMENTS installed sstsimulator.conf file located at $SST_CONFIG_FILE_PATH ==="
-            echo "cat $SST_CONFIG_FILE_PATH"
-            cat $SST_CONFIG_FILE_PATH
+            print_and_dump_loc $SST_CONFIG_FILE_PATH
             echo "=== DONE DUMPING ==="
             echo
 
@@ -2170,15 +2143,13 @@ if [[ $SST_SELECTED_ELEMENTS_CONFIG == "NOBUILD" ]]
 
         echo
         echo "=== DUMPING The SST-EXTERNAL-ELEMENTS installed $HOME/.sst/sstsimulator.conf file ==="
-        echo "cat $HOME/.sst/sstsimulator.conf"
-        cat $HOME/.sst/sstsimulator.conf
+        print_and_dump_loc $HOME/.sst/sstsimulator.conf
         echo "=== DONE DUMPING ==="
         echo
 
         echo
         echo "=== DUMPING The SST-EXTERNAL-ELEMENTS installed sstsimulator.conf file located at $SST_CONFIG_FILE_PATH ==="
-        echo "cat $SST_CONFIG_FILE_PATH"
-        cat $SST_CONFIG_FILE_PATH
+        print_and_dump_loc $SST_CONFIG_FILE_PATH
         echo "=== DONE DUMPING ==="
         echo
 
@@ -2248,15 +2219,13 @@ if [[ $SST_SELECTED_ELEMENTS_CONFIG == "NOBUILD" ]]
 
         echo
         echo "=== DUMPING The JUNO installed $HOME/.sst/sstsimulator.conf file ==="
-        echo "cat $HOME/.sst/sstsimulator.conf"
-        cat $HOME/.sst/sstsimulator.conf
+        print_and_dump_loc $HOME/.sst/sstsimulator.conf
         echo "=== DONE DUMPING ==="
         echo
 
         echo
         echo "=== DUMPING The JUNO installed sstsimulator.conf file located at $SST_CONFIG_FILE_PATH ==="
-        echo "cat $SST_CONFIG_FILE_PATH"
-        cat $SST_CONFIG_FILE_PATH
+        print_and_dump_loc $SST_CONFIG_FILE_PATH
         echo "=== DONE DUMPING ==="
         echo
 
@@ -2435,15 +2404,11 @@ echo "#### FINISHED SETTING UP DIRECTORY STRUCTURE - NOW SETTING ENV RUNTIME VAR
 echo
 echo
 echo
-echo "#### DELETING THE HOME/.sst/sstsimulator.conf file ####"
-echo "#### NOTE: THIS CODE MAY NEED TO BE REMOVED IN THE NEAR FUTURE"
-echo "BEFORE:ls $HOME/.sst/sstsimulator.conf"
-ls $HOME/.sst/sstsimulator.conf
-echo "rm -f $HOME/.sst/sstsimulator.conf"
-rm -f $HOME/.sst/sstsimulator.conf
-echo "AFTER: ls $HOME/.sst/sstsimulator.conf"
-ls $HOME/.sst/sstsimulator.conf
-echo "#### DONE DELETING THE HOME/.sst/sstsimulator.conf file ####"
+if [[ -f $HOME/.sst/sstsimulator.conf ]]; then
+    echo "#### DELETING THE HOME/.sst/sstsimulator.conf file ####"
+    rm -f $HOME/.sst/sstsimulator.conf
+    echo "#### DONE DELETING THE HOME/.sst/sstsimulator.conf file ####"
+fi
 echo
 echo
 echo
@@ -2662,7 +2627,7 @@ else
                                 ## Check that python is an acceptably new version of python
                                 pyver=$(python -V 2>&1 | grep -Po '(?<=Python )(.+)')
                                 pyverX=$(echo "${version//./}")
-                                if [["$pyverX" -lt "350" ]]
+                                if [[ "$pyverX" -lt "350" ]]
                                 then
                                     echo "ERROR: FOUND python but version is TOO OLD (<3.5.0)."
                                     exit 128
@@ -2704,9 +2669,7 @@ else
                 # Check that the default Intel PIN module is available
                 # For Linux = pin/pin-3.26-98960-g1fc9d60e6-gcc-linux
                 #           ModuleEx puts the avail output on Stdout (where it belongs.)
-                ModuleEx avail | egrep -q "pin/pin-3.26"
-                if [ $? == 0 ]
-                then
+                if ModuleEx avail | grep -E -q "pin/pin-3.26"; then
                 # if `pin module is available, use pin/pin-3.26.
                     if [ $kernel != "Darwin" ] ; then
                         echo "USING INTEL PIN ENVIRONMENT MODULE pin-3.26-98690-g1fc9d60e6-gcc-linux"
@@ -2899,7 +2862,7 @@ then
                             echo "**************************************************************************"
                             # WE ARE RUNNING THE FRAMEWORKS ELEMENTS FULL SET OF TESTS AFTER DOTESTS() HAVE RUN
                             cd $SST_ROOT
-                            
+
                             $SST_PYTHON_APP_EXE $SST_TEST_FRAMEWORKS_CORE_APP_EXE $SST_TEST_FRAMEWORKS_PARAMS -z -r $SST_MULTI_RANK_COUNT -t $SST_MULTI_THREAD_COUNT
                             core_frameworks_retval=$?
                             $SST_PYTHON_APP_EXE $SST_TEST_FRAMEWORKS_ELEMENTS_APP_EXE $SST_TEST_FRAMEWORKS_PARAMS -k -z -r $SST_MULTI_RANK_COUNT -t $SST_MULTI_THREAD_COUNT
@@ -2915,7 +2878,7 @@ then
                             if [ $core_frameworks_retval -eq 0 ]; then
                                 retval=$elements_frameworks_retval
                             else
-                                retval = $core_frameworks_retval
+                                retval=$core_frameworks_retval
                             fi
                         fi
                         echo "BAMBOO: Combined Frameworks + dotests retval = $retval"
